Some code cleanup and splitups
